Calories and Nutrients In The Food Items Of Trending Fast Food Joints

Calories and Nutrients In The Food Items Of Trending Fast Food Joints

Usage

fastfoodcalories

Format

A data frame with 515 observations and 28 variables.

...1

A unique identifier for each data entry.

restaurant

The name of the fast-food restaurant.

item

The name of the food item from the restaurant's menu.

calories

The calorie content of the food item.

cal_fat

The calorie content from fat in the food item.

total_fat

The total fat content in the food item.

sat_fat

The saturated fat content in the food item.

trans_fat

The trans fat content in the food item.

cholesterol

The cholesterol content in the food item.

sodium

The sodium content in the food item.

total_carb

The total carbohydrate content in the food item.

fiber

The dietary fiber content in the food item.

sugar

The sugar content in the food item.

protein

The protein content in the food item.

vit_a

The vitamin A content in the food item.

vit_c

The vitamin C content in the food item.

calcium

The calcium content in the food item.

salad

A categorization of whether the food item is a salad or not.

carb_percent

The percentage of total carbohydrates in the food item.

protein_percent

The percentage of protein in the food item.

fiber_percent

The percentage of dietary fiber in the food item.

vit_a_percent

The percentage of vitamin A in the food item.

vit_c_percent

The percentage of vitamin C in the food item.

calcium_percent

The percentage of calcium in the food item.

cholesterol_percent

The percentage of cholesterol in the food item.

sugar_percent

The percentage of sugar in the food item.

sodium_percent

The percentage of sodium in the food item.

sat_fat_percentage

The percentage of saturated fat in the food item.

Source

https://github.com/rfordatascience/tidytuesday/blob/master/data/2018/2018-09-04/fastfood_calories.csv